+++ apps/core/cmd/supervisor/router/subscribe.go
@@ -8,17 +8,15 @@ import (
 	"github.com/labstack/echo/v4"
 )
 
-func nodeSubscribeKey(parentId int64) string {
+func nodeSubscribeTopic(parentId int64) string {
 	return fmt.Sprintf("NODE:%d", parentId)
-}
-
-func processSubscribeKey(processId string) string {
-	return fmt.Sprintf("PROCESS:%s", processId)
 }
 
 func (r *supervisorRouter) handleSubscribeWS(c echo.Context) error {
 	req, res := c.Request(), c.Response()
-	r.requireSession(c)
+	sess := r.requireSession(c)
+	sid := sess.Name() // sid 겠지?
+	// log.Printf("\tSID - %s", sid)
 	ws, err := upgrader.Upgrade(res, req, nil)
 	if err != nil {
 		log.Printf("supervisor: upgrade 실패 %v", err)
@@ -26,13 +24,23 @@ func (r *supervisorRouter) handleSubscribeWS(c echo.Context) error {
 	}
 
 	conn := transport.New(ws)
+	r.browsers.Append(conn, sid)
 
+	// 구독중이던 process
+	processes, err := r.processManager.ListSubscriptions(sid)
+	for _, proc := range processes {
+		if _, ok := r.processManager.Get(proc.Uid); ok {
+			r.subscribeHub.Subscribe(processTopic(proc.Uid), conn)
+		}
+	}
 	//node:{parentId} - 0이면 nil인 node들을 구독한다
-	r.subscribeHub.Subscribe(nodeSubscribeKey(0), conn)
+	r.subscribeHub.Subscribe(nodeSubscribeTopic(0), conn)
 
 	err = conn.Serve()
 	conn.Close(err)
+	// 모든 구독 해제
 	r.subscribeHub.UnsubscribeAll(conn)
+	r.browsers.Remove(conn)
 
 	return nil
 }