+++ apps/frontend/src/common/api/query.util.ts
@@ -2,6 +2,7 @@ import type { Ref } from 'vue'
 
 export type QueryOptions = Partial<{
     enabled: Ref<boolean>
+    refetchInterval: Ref<number>
 }>
 
 export type ToRefs<T> = {